Move Clb module into tests and remove redundant `Address` module
